Move scrapy/contrib/spiders to scrapy/spiders
