+++ scrapy/templates/spiders/csvfeed.tmpl
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from scrapy.contrib.spiders import CSVFeedSpider
+from scrapy.spiders import CSVFeedSpider
 
 from $project_name.items import ${ProjectName}Item
