+++ scrapy/templates/spiders/xmlfeed.tmpl
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from scrapy.contrib.spiders import XMLFeedSpider
+from scrapy.spiders import XMLFeedSpider
 
 from $project_name.items import ${ProjectName}Item
